requirements.txt
flask==2.3.3
flask-login==0.6.2
flask-sqlalchemy==3.0.5
flask-wtf==1.1.1
Pillow==10.0.0
email-validator==2.0.0
flask-migrate==4.0.4
python-dotenv==1.0.0
werkzeug==2.3.7 